Avg_Temperatures_UMAP_KMeans/requirements.txt
numpy
pandas
matplotlib
umap-learn
scikit-learn